Arrumando erros
Essa branch esta sendo criada pois a anterior não estamos conseguindo desfazer os ultimos 4 commits! E precisamos revertelos

+++ Atlantida/src/main/java/com/ariel/Atlantida/dto/ProdutoDtoCreate.java
@@ -13,7 +13,7 @@ public class ProdutoDtoCreate {
     @Column(nullable = false, length = 100)
     private String tipo;
     @Column(nullable = false, precision = 10, scale = 2)
-    private BigDecimal preco;
+    private double preco;
     @Column(nullable = false)
     private Integer estoque;
 
@@ -41,11 +41,11 @@ public class ProdutoDtoCreate {
         this.tipo = tipo;
     }
 
-    public BigDecimal getPreco() {
+    public double getPreco() {
         return preco;
     }
 
-    public void setPreco(BigDecimal preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
